requirement.txt
tqdm
requests
flask
tkinter
Pillow